project/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
apscheduler==3.10.4
requests==2.32.3
pyyaml==6.0.2
pydantic==2.11.1
aiofiles==24.1.0
